fix `uint64_t` vs `unsigned long long int` warning
git-svn-id: http://ezix.org/source/packages/lshw/development@2426 811e2811-9fd9-0310-a116-b6e8ac943c8b

+++ src/core/pci.cc
@@ -1157,7 +1157,7 @@ bool scan_pci(hwNode & n)
             {
               while(!feof(resource))
               {
-                uint64_t start, end, flags;
+                unsigned long long start, end, flags;
 
                 start = end = flags = 0;
